Object files moved to Components folder.
