+++ VSEngine/Entry.cpp
@@ -1,9 +1,11 @@
 #include "../Renderer/Renderer.h"
+#include "../Geometry/Mesh.h"
 
 int main()
 {
 	VSEngine::Renderer renderer(600,800, "My Window");
-	renderer.Start();
+
+	//renderer.Start();
 	
 	return 0;
 }
