+++ VSEngine/VSEngine.vcxproj.filters
@@ -53,7 +53,7 @@
     <ClCompile Include="Entry.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Object.cpp">
+    <ClCompile Include="..\Components\Object.cpp">
       <Filter>Object</Filter>
     </ClCompile>
   </ItemGroup>
@@ -85,7 +85,7 @@
     <ClInclude Include="..\Renderer\Renderer.h">
       <Filter>Renderer</Filter>
     </ClInclude>
-    <ClInclude Include="Object.h">
+    <ClInclude Include="..\Components\Object.h">
       <Filter>Object</Filter>
     </ClInclude>
   </ItemGroup>
